+++ ExposurePlayback/Events/Error.swift
@@ -24,19 +24,19 @@ extension Playback {
         /// Platform-dependent error code
         internal let code: Int
         
-        /// Error Domain
-        internal let domain: String
-        
         /// Additional detailed error information
         internal let info: String?
         
-        internal init(timestamp: Int64, offsetTime: Int64?, message: String, code: Int, domain: String, info: String? = nil) {
+        /// Error Details, such as stack trace or expanded error info
+        internal let details: String?
+        
+        internal init(timestamp: Int64, offsetTime: Int64?, message: String, code: Int, info: String? = nil, details: String? = nil) {
             self.timestamp = timestamp
             self.offsetTime = offsetTime
             self.message = message
             self.code = code
-            self.domain = domain
             self.info = info
+            self.details = details
         }
     }
 }
@@ -55,8 +55,7 @@ extension Playback.Error: AnalyticsEvent {
             JSONKeys.eventType.rawValue: eventType,
             JSONKeys.timestamp.rawValue: timestamp,
             JSONKeys.message.rawValue: message,
-            JSONKeys.code.rawValue: code,
-            JSONKeys.domain.rawValue: domain
+            JSONKeys.code.rawValue: code
         ]
         
         if let offset = offsetTime {
@@ -65,6 +64,10 @@ extension Playback.Error: AnalyticsEvent {
         
         if let info = info {
             json[JSONKeys.info.rawValue] = info
+        }
+        
+        if let details = details {
+            json[JSONKeys.details.rawValue] = details
         }
         
         return json
@@ -76,8 +79,8 @@ extension Playback.Error: AnalyticsEvent {
         case offsetTime = "OffsetTime"
         case message = "Message"
         case code = "Code"
-        case domain = "Domain"
         case info = "Info"
+        case details = "Details"
     }
 }